+++ src/readabs/__init__.py
@@ -2,11 +2,8 @@
 the Australian Bureau of Statistics (ABS) 
 and the Reserve Bank of Australia (RBA)."""
 
-from . import readabs
-from .readabs import *
-
-__version__ = "0.0.15a1"
-
+__version__ = "0.0.15a3"
+__author__ = "Bryan Palmer"
 __all__ = (
     # -- abs -- related
     "metacol",
@@ -32,3 +29,23 @@ __all__ = (
     "recalibrate",
     "recalibrate_value",
 )
+
+
+# local imports
+from readabs.search_abs_meta import search_abs_meta, find_abs_id
+from readabs.abs_catalogue_map import abs_catalogue
+from readabs.read_abs_cat import read_abs_cat
+from readabs.read_abs_series import read_abs_series
+from readabs.grab_abs_url import grab_abs_url
+from readabs.abs_meta_data_support import metacol
+from readabs.get_rba_links import print_rba_catalogue, rba_catalogue
+from readabs.read_rba_table import read_rba_table, rba_metacol, read_rba_ocr
+from readabs.recalibrate import recalibrate, recalibrate_value
+from readabs.utilities import (
+    percent_change,
+    annualise_rates,
+    annualise_percentages,
+    qtly_to_monthly,
+    monthly_to_qtly,
+)
+from readabs.print_abs_catalogue import print_abs_catalogue